First draft of weft. Underbaked.
The biggest issue here is that the API for communicating with the weft involves these discrete event
handlers, and that discourages batching. It doesn't *prevent* it, but it definitely discourages it.

So the thing about Haxl is that you want to ensure that the following code does the right thing,

    let requests = [];
    for (let id of ids) {
        requests.push(this.Things.getByID(id));
    }
    let responses = await Promise.all(requests);

Right now we synchronously pass to the handler which synchronously returns to us a Promise, and the
naive approach in the handler is to immediately launch the query. And a plugin *can* do it, but the
handler has to look like:

    let pluginHandlingThread = null;
    let queue = [];
    function batchSubmit() {
      // batch handle things in the queue
    }
    module.exports.namespace = 'Things';
    module.exports.init = router => {
      router.registerAsync('getByID', id => {
        queue.push(['getByID', id]);
        if (pluginHandlingThread === null) {
          pluginHandlingThread = setTimeout(batchSubmit, 1000);
        }
      });
    };

We ideally want to put this boilerplate into the handler by default.